Extract SkippyOutput
